// front end/prodesp-gsnet-monitor/src/lib/src/models/historicojustificativa.ts
export interface HistoricoJustificativa {
    JustificativaFME:        string;
    JustificativaCompras:     string;
    JustificativaCAF:         string;
    JustificativaCAFPublica:  string;
    SaldoDisponivel:          string;
    DataUltimaFatura:         string;
    DataUltimoEmpenho:        string;
    DataMonitoramento:        string;
    DataInicioVigencia:       string;
    DataMonitoramentoExtenso: string;
    ConsumoMedex:             string;
}